+++ scorex-basics/src/main/scala/scorex/api/http/PeersApiRoute.scala
@@ -4,7 +4,6 @@ import java.net.{InetAddress, InetSocketAddress}
 import javax.ws.rs.Path
 
 import akka.actor.ActorRefFactory
-import akka.http.scaladsl.model.{ContentTypes, HttpEntity}
 import akka.http.scaladsl.server.Route
 import akka.pattern.ask
 import io.swagger.annotations._
@@ -15,6 +14,7 @@ import scorex.network.NetworkController.ConnectTo
 import scorex.network.peer.{PeerInfo, PeerManager}
 
 import scala.concurrent.ExecutionContext.Implicits.global
+import scala.util.Try
 
 @Path("/peers")
 @Api(value = "/peers", description = "Get info about peers", position = 2)
@@ -72,19 +72,26 @@ case class PeersApiRoute(override val application: Application)(implicit val con
   @Path("/connect")
   @ApiOperation(value = "Connect to peer", notes = "Connect to peer", httpMethod = "POST")
   @ApiImplicitParams(Array(
-    new ApiImplicitParam(name = "host", value = "Peer host", required = true, paramType = "body", dataType = "String"),
-    new ApiImplicitParam(name = "port", value = "Peer port", required = true, paramType = "body", dataType = "Int")
-  ))
-  def connect: Route = path("connect") {
-    val host: String = "127.0.0.1"
-    val port: Int = 9084
-
-    withCors {
-      entity(as[String]) { body =>
-        val add: InetSocketAddress = new InetSocketAddress(InetAddress.getByName(host), port)
-        application.networkController ! ConnectTo(add)
-        val resp = Json.obj(add.getHostName -> "Trying to connect")
-        complete(HttpEntity(ContentTypes.`application/json`, resp.toString()))
+    new ApiImplicitParam(
+      name = "body",
+      value = "Json with data",
+      required = true,
+      paramType = "body",
+      defaultValue = "{\n\t\"host\":\"127.0.0.1\",\n\t\"port\":\"9084\"\n}"
+    )
+  )) def connect: Route = path("connect") {
+    entity(as[String]) { body =>
+      withAuth {
+        postJsonRoute {
+          Try {
+            val js = Json.parse(body)
+            val host = (js \ "host").as[String]
+            val port = (js \ "port").as[Int]
+            val add: InetSocketAddress = new InetSocketAddress(InetAddress.getByName(host), port)
+            application.networkController ! ConnectTo(add)
+            Json.obj("hostname" -> add.getHostName, "status" -> "Trying to connect")
+          }.getOrElse(WrongJson.json)
+        }
       }
     }
   }