+++ scorex-transaction/src/main/scala/scorex/api/http/PaymentApiRoute.scala
@@ -80,7 +80,5 @@ case class PaymentApiRoute(override val application: Application)(implicit val c
         }
       }
     }
-
-
   }
 }